+++ src/graphlab/util/stl_util.hpp
@@ -413,7 +413,7 @@ namespace graphlab {
                                 const size_t npad,
                                 const char pad_value = '0') {
     std::stringstream strm;
-    strm << std::setw(npad) << std::setfill(pad_value)
+    strm << std::setw((int)npad) << std::setfill(pad_value)
          << number;
     return strm.str();
   }
